+++ packages/score-format/src/index.ts
@@ -2,6 +2,24 @@ export {
   describeScore,
   FORMAT_VERSION,
   isSupportedVersion,
+  timingOf,
   type Score,
   type ScoreMetadata,
 } from './score'
+
+export {
+  barAtTick,
+  barRangeToTicks,
+  DEFAULT_MICROSECONDS_PER_QUARTER,
+  DEFAULT_TICKS_PER_QUARTER,
+  resolveTiming,
+  secondsToTicks,
+  tickAtBar,
+  ticksPerBar,
+  ticksToSeconds,
+  type BarPosition,
+  type ResolvedTiming,
+  type TempoEvent,
+  type TimeSignatureEvent,
+  type Timing,
+} from './time'